refactor: remove channels where unnecessary
Instead of having an architecture where the spans come out of the
receiver from a channel and then we process the same trace several times
by several "middleware components" such as CutoffFilter,
SublayerTagger... by sending the trace on an `in` channel and then
waiting for the result on an `out` channel,
we just make the middleware a function that we can call on the main
loop.
It's also good for tests and benchmarks that become much simpler to
write and expect.

So far the result seems promising for something as dumb as the sublayer
tagger we just get a free boost (the allocs come from an unrelated
optimization, pre-allocating the result):
```
➜  raclette git:(master) ✗ benchcmp model/sublayers_pre_refactor.txt
model/sublayers_post_refactor.txt
benchmark                   old ns/op     new ns/op     delta
BenchmarkSublayerThru-4     5233          4014          -23.29%

benchmark                   old allocs     new allocs     delta
BenchmarkSublayerThru-4     27             17             -37.04%

benchmark                   old bytes     new bytes     delta
BenchmarkSublayerThru-4     2240          1696          -24.29%

```